requirements.txt
flask
pymysql
flask-sqlalchemy
sqlalchemy
python-dotenv
google-cloud-firestore